[#651] Rename examples packages
To `org.hibernate.reactive.example.*`
