+++ examples/native-sql-example/src/main/java/org/hibernate/reactive/example/nativesql/Author.java
@@ -1,19 +1,24 @@
-package org.hibernate.example.session.reactive;
+package org.hibernate.reactive.example.nativesql;
+
+import org.hibernate.annotations.SQLDelete;
+import org.hibernate.annotations.SQLInsert;
+import org.hibernate.annotations.SQLUpdate;
+import org.hibernate.annotations.Subselect;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
-import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.ArrayList;
 import java.util.List;
 
-import static javax.persistence.CascadeType.PERSIST;
-
 @Entity
-@Table(name="authors")
+@Subselect("select name, id from authors")
+@SQLInsert(sql = "insert into authors (name, id) values ($1, $2)")
+@SQLUpdate(sql = "")
+@SQLDelete(sql = "delete from authors where id = $1")
 class Author {
 	@Id @GeneratedValue
 	private Integer id;
@@ -21,7 +26,7 @@ class Author {
 	@NotNull @Size(max=100)
 	private String name;
 
-	@OneToMany(mappedBy = "author", cascade = PERSIST)
+	@OneToMany(mappedBy = "author")
 	private List<Book> books = new ArrayList<>();
 
 	Author(String name) {
