+++ examples/session-example/build.gradle
@@ -68,7 +68,7 @@ dbs.each { db ->
                 group = "Execution"
                 description = "Run ${mainClass} on ${db}"
                 classpath = sourceSets.main.runtimeClasspath
-                main = "org.hibernate.example.session.reactive.${mainClass}"
+                main = "org.hibernate.reactive.example.session.${mainClass}"
                 // The persistence unit name defined in resources/META-INF/persistence.xml
                 args db.toLowerCase() + '-example'
             }
