+++ examples/session-example/src/main/resources/META-INF/persistence.xml
@@ -6,8 +6,8 @@
     <persistence-unit name="postgresql-example">
         <provider>org.hibernate.reactive.provider.ReactivePersistenceProvider</provider>
 
-        <class>org.hibernate.example.session.reactive.Author</class>
-        <class>org.hibernate.example.session.reactive.Book</class>
+        <class>org.hibernate.reactive.example.session.Author</class>
+        <class>org.hibernate.reactive.example.session.Book</class>
 
         <properties>
 
@@ -41,8 +41,8 @@
     <persistence-unit name="mysql-example">
         <provider>org.hibernate.reactive.provider.ReactivePersistenceProvider</provider>
 
-        <class>org.hibernate.example.session.reactive.Author</class>
-        <class>org.hibernate.example.session.reactive.Book</class>
+        <class>org.hibernate.reactive.example.session.Author</class>
+        <class>org.hibernate.reactive.example.session.Book</class>
 
         <properties>
